Heimdal: Undo changes since 2017-09
This will allow us to import lorikeet-heimdal-201709200427 (commit 9b4808577ad0f4a3a4a55a3dc32d6739e387135b)
and then to bring these back via a fresh import of a current branch.

+++ source4/heimdal/lib/asn1/gen.c
@@ -502,7 +502,7 @@ is_primitive_type(int type)
 }
 
 static void
-space(size_t level)
+space(int level)
 {
     while(level-- > 0)
 	fprintf(headerfile, "  ");
@@ -531,7 +531,7 @@ have_ellipsis(Type *t)
 }
 
 static void
-define_asn1 (size_t level, Type *t)
+define_asn1 (int level, Type *t)
 {
     switch (t->type) {
     case TType:
